fix: correct time calculation and coordinate parsing bugs
- Fix `calculateTimeBetween` divisor to use milliseconds (1000) instead of seconds (60)
- Change coordinate parsing from `parseInt` to `parseFloat` for decimal values
- Add new `format-time` helper with `formatTimeMs` and `formatTimeSeconds` functions
- Replace inline time formatting with shared helper functions across components
- Fix date formatting bug where `getDay()` was incorrectly used instead of `getDate()`
- Improve form validation in update user form to include website field
- Remove unused keyboard animation code and fix TypeScript generics in selectors
- Add proper error handling and success messages for form submissions
- Update component name from `UpdateUserForm` to `UpdateClientParkingForm`

+++ apps/mobile/src/lib/helpers/get-coordinates-from-url.ts
@@ -24,8 +24,8 @@ export const extractCoordinates = async (shortUrl: string) => {
     }
 
     if (match) {
-      const latitude = parseInt(match[1]);
-      const longitude = parseInt(match[2]);
+      const latitude = parseFloat(match[1]);
+      const longitude = parseFloat(match[2]);
       return { latitude, longitude };
     } else {
       console.log('Coordinates not found in the URL');
